test: reduce importOriginal in slow unit tests
Amp-Thread-ID: https://ampcode.com/threads/T-01a068ec-20b6-7608-be1c-d7fd49450fd9
